requirements.txt
django>=4.2
markdown
sentence-transformers
faiss-cpu
numpy
pandas
requests
bcrypt
